strongly type graphql queries
